+++ CHANGELOG.md
@@ -12,5 +12,7 @@ This document contains a list of all the changes for each version of Cortex.
 
 ## Version 1.1
 
+*Release date to be determined*
+
 - Reduced default heartbeat interval.
 - Updated image preview to support modern Predator.