+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This document contains a list of all the changes for each version of Cortex.
 
 ## Version 1.1
 
-*Release date to be determined*
+December 5th, 2023
 
 - Reduced default heartbeat interval.
 - Updated image preview to support Predator version 9.0.